ADdin

+++ backend/routes/productOrders.js
@@ -1,5 +1,5 @@
 const express = require('express');
-const { addProductOrder, deleteProductOrder, deleteAllUserProductOrders, getAllProductOrders, getUserProductOrders, updateProductOrder } = require('../controllers/productsOrders');
+const { addProductOrder, deleteProductOrder, deleteAllUserProductOrders, getAllProductOrders, getUserProductOrders, updateProductOrder, getAllProductOrdersForAdmin } = require('../controllers/productsOrders');
 const authentication = require("../middlewares/authentication");
 const authorization = require("../middlewares/authorization")
 const productOrderRouter = express.Router();
@@ -10,6 +10,8 @@ productOrderRouter.delete('/all/order',  authentication, authorization("DELETE_P
 productOrderRouter.get('/', getAllProductOrders)
 productOrderRouter.get('/showCart', authentication, getUserProductOrders)
 productOrderRouter.put('/:id', authentication, authorization("UPDATE_PRODUCT_ORDER"), updateProductOrder)
+productOrderRouter.get('/all/admin', getAllProductOrdersForAdmin)
+
 
 module.exports = productOrderRouter;
